requirements.txt
python-telegram-bot[socks]==21.7
jieba==0.42.1
cryptography==44.0.0
bcrypt==4.2.1
pytest==8.3.4

